Fix detection of closed connections
When the remote peer closes a connection, this causes the rx task to
stop, but the tx side keeps running until it is used to send a message,
which is guaranteed to fail. The fix is to add a 'closed' signal to
tell the tx task to stop after the rx task finishes.